# Ensure that bitwise operations of small masks can still be folded