This resolves an issue spotted in https://github.com/dotnet/runtime/pull/117865#discussion_r2219620345