@tannergooding 2025-07-21
For example, I get the following numbers on my 7900X

### Before

| Method              | Mean      | Error     | StdDev    | Median    | Min       | Max       | Allocated |
|-------------------- |----------:|----------:|----------:|----------:|----------:|----------:|----------:|
| Throughput          |  3.174 us | 0.0816 us | 0.0940 us |  3.135 us |  3.063 us |  3.315 us |         - |
| SV_Throughput       |  3.967 us | 0.0220 us | 0.0172 us |  3.964 us |  3.947 us |  4.004 us |         - |
| SV_ThroughputIC     |  4.519 us | 0.0854 us | 0.0799 us |  4.549 us |  4.431 us |  4.614 us |         - |
| FalsePositives      | 11.598 us | 0.1648 us | 0.1542 us | 11.598 us | 11.383 us | 11.819 us |         - |
| SV_FalsePositives   |  9.393 us | 0.1852 us | 0.1819 us |  9.468 us |  9.058 us |  9.696 us |         - |
| SV_FalsePositivesIC | 10.724 us | 0.1985 us | 0.1857 us | 10.705 us | 10.507 us | 10.984 us |         - |

### After

| Method              | Mean      | Error     | StdDev    | Median    | Min       | Max       | Allocated |
|-------------------- |----------:|----------:|----------:|----------:|----------:|----------:|----------:|
| Throughput          |  3.042 us | 0.0342 us | 0.0303 us |  3.032 us |  3.005 us |  3.105 us |         - |
| SV_Throughput       |  3.956 us | 0.0058 us | 0.0045 us |  3.956 us |  4.247 us |  4.262 us |         - |
| SV_ThroughputIC     |  4.647 us | 0.0276 us | 0.0215 us |  4.651 us |  4.611 us |  4.682 us |         - |
| FalsePositives      | 11.619 us | 0.2319 us | 0.2169 us | 11.543 us | 11.327 us | 12.075 us |         - |
| SV_FalsePositives   |  9.513 us | 0.1498 us | 0.1328 us |  9.518 us |  9.356 us |  9.758 us |         - |
| SV_FalsePositivesIC | 10.930 us | 0.0526 us | 0.0439 us | 10.928 us | 10.874 us | 11.026 us |         - |

Different boxes report different numbers, but all are generally inline with the change being the same or faster and with smaller average codegen. The 7900X remains slightly pessimized for V512 related mask usage due to its double pumping, which makes the mask<->vector conversions a bit more visible.