@MihaZupan 2025-07-21
Here are 7 consecutive runs from my 9950x consistently showing a similar regression as reported by the bot for the Epyc 9V74.
https://gist.github.com/MihaZupan/9cfd8e479bb95e8cd174e9704d6731ab

Or another run with `LongJob` (so multiple restarts) on the same Zen 4 Epyc: https://github.com/MihuBot/runtime-utils/issues/1252