@tannergooding 2025-07-21
> Here are 7 consecutive runs from my 9950x consistently showing a similar regression as reported by the bot for the Epyc 9V74.

I'd speculate there's something else at play here in the microbenchmarks then.

I still cannot reproduce locally (Tiger Lake, Cascade Lake, Ice Lake, or Zen4) and this is a pure reduction in terms of uops, instructions, and code size. It follows the recommendations from the Intel/AMD optimization manuals and more closely matches what GCC/Clang/MSVC produce for similar code.