@tannergooding 2025-07-21
> more closely matches what GCC/Clang/MSVC produce for similar code.

What's lacking is still the handling of `vpmovm2*; vptest` and similar; which is a bit of a more involved change as indicated above.