@tannergooding 2025-07-22
@EgorBo numbers for sapphire rapids/cascade lake look like what I saw, and what I see on my Zen4 and Tiger Lake boxes (minus the outlier you have for `FalsePositives` on cascade lake).

I think this is still something we should take. This is inline with the intended optimization the path should have already been doing, is an instruction count reduction, a micro-op reduction, a (as computed by LLVM MCA) latency reduction, a reduction in assembly size, and matches the general optimization guidelines around AVX512 to avoid unnecessary conversions between mask and vector form.

Even if there is a regression on Zen5, this would be micro-architecture specific and something that should get addressed over time, particularly as we finish the other remaining optimization work to ensure we consume the kmask directly in the subsequent branch tests (i.e. we generate `kortest` directly, rather than `vpmovm2*; vptest`)